refactor(rules): promote rules/mod.rs to a directory
Splits the 2224-line rules/mod.rs into focused submodules:

- types.rs       — Rule, PackEntry, GateFailure, RuleMatch
- pattern.rs     — CompiledRule + compile/match helpers (private)
- grouping.rs    — group_by_handler, handler_execution_order
- scanner.rs     — Scanner struct + walk/match impl + SPECIAL_FILES,
                   should_skip_entry, gate evaluation
- mod.rs         — re-exports the public surface

Tests are co-located with the code they cover: pattern unit tests in
pattern.rs, grouping tests in grouping.rs, serialisation in types.rs,
and the full scanner / gate / C2 / C4 integration suite in scanner.rs.
Public API (Rule, Scanner, PackEntry, GateFailure, RuleMatch,
should_skip_entry, group_by_handler, handler_execution_order,
SPECIAL_FILES) is preserved via re-exports.
